[toa] Add a support of building for the Android.
As we make a build for the Android we have to use android toolchain.

Way we recompile an android toolchain sources by gbs and use result
rpm as no-cross-compile toolchain is a little bit complex, at least now.
So we use x86_64 android toolchain as a cross-compile toolchain.

Look to how_to_deploy.txt for an explanations.
Later we can migrate to the gbs for building for the Android.

Change-Id: I0b12def8d809de0335c1b97aa0073d431677c605
Signed-off-by: Sergey Sizonov <s.sizonov@samsung.com>