[toa] Add a support of building for the Android.
As we make a build for the Android we have to use android toolchain.

Way we recompile an android toolchain sources by gbs and use result
rpm as no-cross-compile toolchain is a little bit complex, at least now.
So we use x86_64 android toolchain as a cross-compile toolchain.

Look to how_to_deploy.txt for an explanations.
Later we can migrate to the gbs for building for the Android.

Change-Id: I4149564af164b2a553f0d01791b9a05a6b2ce29d
Signed-off-by: Sergey Sizonov <s.sizonov@samsung.com>